pyproject.toml
[tool.poetry]
name = "enunlg"
version = "0.1.0"
description = "Edinburgh Napier University package for Natural Language Generation"
authors = ["David M. Howcroft <dave.howcroft@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
regex = "^2021.4.4"
numpy = "^1.20.3"
bidict = "^0.21.2"
python-box = "^5.3.0"
seaborn = "^0.12.2"
jupyter = "^1.0.0"
ipython = "^7.25.0"
ipymd = "^0.1.3"
nbstripout = "^0.5.0"
# Uncomment the torch line relevant to your environment!
# Live dangerously and let poetry try to find something; this didn't work for me
#torch = "^1.9.0+cpu"
# Torch 1.9.1 on CPU-only for Python 3.9 on 64-bit Linux
torch = {url="https://download.pytorch.org/whl/cpu/torch-1.9.1%2Bcpu-cp39-cp39-linux_x86_64.whl"}
# Torch 1.9.1 on CPU-only for Python 3.9 on 64-bit Windows
#torch = {url="https://download.pytorch.org/whl/cpu/torch-1.9.1%2Bcpu-cp39-cp39-win_amd64.whl"}
# Torch 1.9.1 on CPU-only for Python 3.9 on 64-bit Intel CPU MacOSX 10
#torch = {url="https://download.pytorch.org/whl/cpu/torch-1.9.1-cp39-none-macosx_10_9_x86_64.whl"}
# Torch 1.9.1 on CPU-only for Python 3.9 on 64-bit ARM MacOSX 11
#torch = {url="https://download.pytorch.org/whl/cpu/torch-1.9.1-cp39-none-macosx_11_0_arm64.whl"}
# Torch 1.9.1 on CPU-only for Python 3.9 on 64-bit Linux with CUDA 11
#torch = {url="https://download.pytorch.org/whl/cu111/torch-1.9.0%2Bcu111-cp39-cp39-linux_x86_64.whl"}
hydra-core = "^1.1.1"
pytorch-lightning = "^1.5.4"
prettytable = "^3.4.1"
matplotlib = "^3.6.2"
scikit-image = "^0.19.3"
future = "^0.18.2"
pytorch-beam-search = "^1.2.2"
sacrebleu = "^2.3.1"
scalene = "^1.5.13"
lxml = "^4.9.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
